+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- deps: lazy extension dependencies.
+- feat: markdown linting via `markdownlint`.
+- feat: lazy extension dependencies.
 
 ## 0.11.0 (2025-02-20)